+++ Changes.md
@@ -9,6 +9,8 @@ developed by
 
 ## New features
 
+* NEW Dpdl language plug-in 'Wgsl' that allows to scale computations on GPU's by using WGSL code (WebGPU Shading Language)
+
 * NEW Dpdl Wasm runtime plug-in for compiling and accessing 'Wasm' module functions from 'dpdl' and from embedded code
 
 * Multi-line structured text, data and code resources are supported